Dual Encryption/requirements.txt
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.9.0.80
